+++ CHANGES.rst
@@ -13,6 +13,7 @@
 - Sideview: minor changes in look&feel (#891)
 - Sideview: current route is now kept when maptool is deselected (#891)
 - Graph: added batch delete option (#884)
+- Graph: features with same id, but different layer should have different color (#895)
 
 
 3.1.8 (2023-09-04)